Adding dependency in pom.xml for swagger

+++ src/main/java/com/techreturners/bookmanager/controller/BookManagerController.java
@@ -1,6 +1,6 @@
 package com.techreturners.bookmanager.controller;
 
-import com.techreturners.bookmanager.model.Book;
+import com.techreturners.bookmanager.model.*;
 import com.techreturners.bookmanager.service.BookManagerService;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -8,6 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/api/v1/book")
@@ -20,13 +21,21 @@ public class BookManagerController {
     }
 
     @GetMapping
-    public ResponseEntity<List<Book>> getAllBooks() {
+    public ResponseEntity<?> getAllBooks() {
         List<Book> books = bookManagerService.getAllBooks();
+        if(books.isEmpty())
+             {
+                ErrorMessage errorResponse = new ErrorMessage();
+                errorResponse.setMessage("Record not found");
+                return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+            }
+        else
         return new ResponseEntity<>(books, HttpStatus.OK);
     }
 
     @GetMapping({"/{bookId}"})
     public ResponseEntity<Book> getBookById(@PathVariable Long bookId) {
+        if(Objects.isNull(bookId)) throw new NullPointerException();
         return new ResponseEntity<>(bookManagerService.getBookById(bookId), HttpStatus.OK);
     }
 
@@ -41,12 +50,14 @@ public class BookManagerController {
     //User Story 4 - Update Book By Id Solution
     @PutMapping({"/{bookId}"})
     public ResponseEntity<Book> updateBookById(@PathVariable("bookId") Long bookId, @RequestBody Book book) {
+        if(Objects.isNull(bookId)) throw new NullPointerException();
         bookManagerService.updateBookById(bookId, book);
         return new ResponseEntity<>(bookManagerService.getBookById(bookId), HttpStatus.OK);
     }
 
     @DeleteMapping({"/{bookId}"})
     public ResponseEntity<List<Book>> deleteBookById(@PathVariable Long bookId ){
+        if(Objects.isNull(bookId)) throw new NullPointerException();
         List<Book> books =bookManagerService.deleteBookById(bookId);
         return new ResponseEntity<>(books, HttpStatus.OK);
     }
